% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RocketFunctions_Laureen Dunker.R
\name{getRocketData}
\alias{getRocketData}
\title{getRocketData}
\usage{
getRocketData()
}
\arguments{
\item{none}{no parameters needed}
}
\value{
Returns a concise data frame with all  rocket launches in 2024 and additional information about launches.
}
\description{
Get the current data from Launch Library 2 API, which is needed to run the Rocket Launches Shiny App.
}
\note{
Packages needed: httr, jsonlite, dplyr; If warning is issued regarding GET function, run getRocketData() again until the data is returned.
}
\examples{
# data_rockets <- getRocketData()
}
